# Improve branch misses on frozen object predicate checks negatively affecting performance of most setters as most objects are not frozen